made follow page and created hidden email-sub-management to make sub/unsub easier
