+++ mobile-web-app/src/css/EmailSubscriptionManagement.css
@@ -42,26 +42,3 @@
     border: 1px solid grey;
     background-color: #d1e0e0;
 }
-
-.follow-the-blokes-text {
-    color: white;
-    font-weight: bolder;
-    font-size: 40px;
-    padding: 10px;
-}
-
-.head {
-    width: 150px;
-    height: auto;
-}
-
-.bloke-name {
-    font-size: 35px;
-    color: white;
-}
-
-.social-media-logo{
-    width: 30px;
-    height: auto;
-    margin: 5px;
-}
